@@ -1,11 +1,12 @@
 {
   "name": "@dan-schel/js-utils",
-  "version": "3.3.6",
+  "version": "3.3.7",
   "description": "A library for code common to my various NodeJS/Browser projects.",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
   "umd:main": "dist/umd/index.js",
   "types": "dist/types/index.d.ts",
+  "type": "module",
   "scripts": {
     "clean": "rimraf dist",
     "prebuild": "npm run clean",
@@ -15,8 +16,8 @@
     "build:umd": "webpack --config config/webpack.config.js",
     "build:types": "tsc -p config/tsconfig.types.json",
     "package": "npm run build && npm pack",
-    "test": "jest",
-    "test-coverage": "jest --coverage",
+    "test": "vitest --run",
+    "test-coverage": "vitest --run --coverage",
     "lint": "eslint . --max-warnings 0",
     "format": "prettier --write .",
     "format-check": "prettier --check ."
@@ -45,19 +46,18 @@
     "uuid": "9.0.1"
   },
   "devDependencies": {
-    "@types/jest": "29.5.14",
+    "@eslint/js": "9.25.1",
     "@types/uuid": "10.0.0",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
-    "eslint": "7.32.0",
+    "@vitest/coverage-v8": "^3.1.2",
+    "eslint": "9.25.1",
     "eslint-config-prettier": "10.1.2",
-    "eslint-plugin-prettier": "4.2.1",
-    "jest": "29.7.0",
-    "prettier": "2.8.8",
+    "eslint-plugin-prettier": "5.2.6",
+    "prettier": "3.5.3",
     "rimraf": "6.0.1",
-    "ts-jest": "29.3.2",
     "ts-loader": "9.5.2",
     "typescript": "5.8.3",
+    "typescript-eslint": "8.31.0",
+    "vitest": "^3.1.2",
     "webpack": "5.99.6",
     "webpack-cli": "6.0.1"
   }
